Move remainder of Pwm wrapper into Motor.
